chore(types): update types
