+++ packages/serendipity-core/src/plugin-wrapper.ts
@@ -12,9 +12,9 @@ import {
   PLUGIN_INQUIRY_META_KEY,
   PLUGIN_NAME_META_KEY, PLUGIN_RUNTIME_META_KEY, PLUGIN_SCRIPT_META_KEY
 } from './common'
-import { PluginModuleInfo } from './types/plugin'
 import { PluginMethodMetaBase } from './types/pluginMeta'
 import 'reflect-metadata'
+import { PluginModuleInfo } from './types/common'
 
 
 export class PluginWrapper {
